fix: tests + feature permutations
